Refactor cache into util sub pkg
